requirements.txt
fastapi
uvicorn
flask
langchain
langchain-community
langchain-openai
langchain-qdrant
langchain-ollama
numpy
openai-whisper
requests
torch
pydantic
python-multipart
qdrant-client
jq
redis
pypdf
unstructured
python-docx
rank_bm25
sentence-transformers
scikit-learn
streamlit
google-generativeai
llama-index
transformers